agentprovision/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0

# Testing
pytest
pytest-cov
pytest-asyncio>=0.18.0,<0.19.0

# Environment
python-dotenv>=0.21.0

# HTTP Client
requests>=2.26.0,<3.0.0
aiohttp>=3.8.0,<4.0.0
httpx>=0.20.0 # Added for API testing

# AI/ML
google-generativeai>=0.3.0
anthropic>=0.7.0
openai>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
email-validator>=2.0.0
bcrypt~=3.2.0

# Code Quality
black
isort
flake8
mypy

# Database
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
asyncpg

# Caching
redis>=4.0.0,<5.0.0
aioredis>=2.0.0,<3.0.0

# File operations
aiofiles>=23.0.0,<24.0.0

# Monitoring
prometheus-client>=0.12.0,<0.13.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-prometheus==0.45b0

# Task Queue
celery>=5.2.0,<6.0.0
flower>=1.0.0,<2.0.0

# Additional
python-dateutil>=2.8.0,<3.0.0
autopep8
psutil>=5.9.0,<6.0.0
